requirements.txt
tensorflow-gpu
keras
h5py
numpy
python-chess